+++ eid-provider-library/HttpRequest.cs
@@ -153,6 +153,8 @@ namespace com.sorlov.eidprovider
                     return false;
                 }
 
+ //dotNet 5.0 offers a nicer approach so if we are building modern then use that custom store otherwise just loop and hope the root is the first in the chain. =)
+#if NET_50
                 for (int i = 1; i < chain.ChainElements.Count; i++)
                 {
                     chain.ChainPolicy.ExtraStore.Add(chain.ChainElements[i].Certificate);
@@ -167,6 +169,10 @@ namespace com.sorlov.eidprovider
                 chain.ChainPolicy.TrustMode = X509ChainTrustMode.CustomRootTrust;
                 chain.ChainPolicy.CustomTrustStore.AddRange(roots);
                 return chain.Build((X509Certificate2)serverCert);
+#else
+                X509Certificate2 chainRoot = chain.ChainElements[chain.ChainElements.Count - 1].Certificate;
+                return roots.Contains(chainRoot);
+#endif
             };
         }